+++ backend/spec/src/main/java/com/sk/skala/axcalibur/spec/feature/report/service/ReportService.java
@@ -2,150 +2,29 @@ package com.sk.skala.axcalibur.spec.feature.report.service;
 
 import com.sk.skala.axcalibur.spec.feature.report.dto.response.TestCaseReportResponseDto;
 import com.sk.skala.axcalibur.spec.feature.report.dto.response.TestScenarioReportResponseDto;
-import com.sk.skala.axcalibur.spec.feature.report.entity.MappingEntity;
-import com.sk.skala.axcalibur.spec.feature.report.entity.ScenarioEntity;
-import com.sk.skala.axcalibur.spec.feature.report.entity.TestCaseDataEntity;
-import com.sk.skala.axcalibur.spec.feature.report.entity.TestCaseEntity;
-import com.sk.skala.axcalibur.spec.feature.report.entity.TestcaseResultEntity;
-import com.sk.skala.axcalibur.spec.feature.report.repository.MappingRepository;
-import com.sk.skala.axcalibur.spec.feature.report.repository.ScenarioRepository;
-import com.sk.skala.axcalibur.spec.feature.report.repository.TestCaseDataRepository;
-import com.sk.skala.axcalibur.spec.feature.report.repository.TestCaseRepository;
-import com.sk.skala.axcalibur.spec.feature.report.repository.TestcaseResultRepository;
-import com.sk.skala.axcalibur.spec.global.code.ErrorCode;
-import com.sk.skala.axcalibur.spec.global.entity.AvalonCookieEntity;
-import com.sk.skala.axcalibur.spec.global.exception.BusinessExceptionHandler;
-import com.sk.skala.axcalibur.spec.global.repository.AvalonCookieRepository;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class ReportService {
-
-    private final ScenarioRepository scenarioRepository;
-    private final MappingRepository mappingRepository;
-    private final TestCaseRepository testCaseRepository;
-    private final TestCaseDataRepository testCaseDataRepository;
-    private final AvalonCookieRepository avalonCookieRepository;
-    private final TestcaseResultRepository testcaseResultRepository;
-    private final FileTemplateService fileTemplateService;
-
+/**
+ * 리포트 관리 서비스 인터페이스
+ * 테스트 시나리오 및 테스트케이스 리포트 생성 및 다운로드 기능 제공
+ */
+public interface ReportService {
+    
     /**
      * 테스트시나리오 리포트 다운로드
+     * @param avalon 사용자 토큰
+     * @return 테스트시나리오 리포트 응답 DTO
+     * @throws IOException 파일 처리 중 오류가 발생한 경우
      */
-    public TestScenarioReportResponseDto downloadTestScenarioReport(String avalon) throws IOException {
-        // 1. 토큰 검증
-        AvalonCookieEntity userInfo = validateAvalonToken(avalon);
-
-        // 2. 시나리오 데이터 조회
-        List<ScenarioEntity> scenarios = scenarioRepository.findByProject_Key(userInfo.getProjectKey());
-        log.info("사용자 시나리오 조회 완료: {}개", scenarios.size());
-
-        // 3. 엑셀 리포트 생성
-        byte[] fileData = fileTemplateService.generateScenarioReport(scenarios);
-        
-        // 4. 응답 반환
-        return TestScenarioReportResponseDto.builder()
-                .avalon(avalon)
-                .fileName("scenario_report.xlsx")
-                .contentType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")
-                .fileData(fileData)
-                .build();
-    }
+    TestScenarioReportResponseDto downloadTestScenarioReport(String avalon) throws IOException;
 
     /**
      * 테스트케이스 리포트 다운로드
+     * @param scenarioId 시나리오 ID
+     * @param avalon 사용자 토큰
+     * @return 테스트케이스 리포트 응답 DTO
+     * @throws IOException 파일 처리 중 오류가 발생한 경우
      */
-    public TestCaseReportResponseDto downloadTestCaseReport(String scenarioId, String avalon) throws IOException {
-        // 1. 토큰 검증
-        validateAvalonToken(avalon);
-
-        // 2. 시나리오 조회
-        ScenarioEntity scenario = getScenario(scenarioId);
-        
-        // 3. 테스트케이스 데이터 조회
-        List<MappingEntity> mappings = getMappingsByScenarioId(scenarioId);
-        List<TestCaseEntity> testCases = getTestCasesByMappings(mappings);
-        List<TestCaseDataEntity> testCaseData = getTestCaseDataByTestCases(testCases);
-        
-        log.info("시나리오 ID: {} 에 대한 테스트케이스 데이터 조회 완료 - 매핑: {}개, 테스트케이스: {}개, 데이터: {}개", 
-                scenarioId, mappings.size(), testCases.size(), testCaseData.size());
-
-        // 4. 테스트케이스 결과 조회 (새로 추가)
-        List<TestcaseResultEntity> testCaseResults = new ArrayList<>();
-        testCases.forEach(testCase -> {
-            List<TestcaseResultEntity> results = testcaseResultRepository.findAllByTestcaseId(testCase.getId());
-            testCaseResults.addAll(results);
-        });
-
-        // 4. 엑셀 리포트 생성
-        byte[] fileData = fileTemplateService.generateTestCaseReport(testCases, testCaseData, testCaseResults);
-        
-        // 5. 응답 반환
-        return TestCaseReportResponseDto.builder()
-                .avalon(avalon)
-                .fileName("testcase_report_" + scenarioId + ".xlsx")
-                .contentType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")
-                .fileData(fileData)
-                .build();
-    }
-
-    /**
-     * 토큰 검증 및 사용자 정보 반환
-     */
-    private AvalonCookieEntity validateAvalonToken(String avalon) {
-        return avalonCookieRepository.findByToken(avalon)
-                .orElseThrow(() -> BusinessExceptionHandler.builder()
-                        .errorCode(ErrorCode.NOT_VALID_COOKIE_ERROR)
-                        .message("토큰이 유효하지 않습니다.")
-                        .build());
-    }
-
-    /**
-     * 시나리오 조회
-     */
-    private ScenarioEntity getScenario(String scenarioId) {
-        return scenarioRepository.findByScenarioId(scenarioId)
-                .orElseThrow(() -> BusinessExceptionHandler.builder()
-                        .errorCode(ErrorCode.NOT_FOUND_ERROR)
-                        .message("시나리오를 찾을 수 없습니다.")
-                        .build());
-    }
-
-    /**
-     * 시나리오별 매핑 조회
-     */
-    private List<MappingEntity> getMappingsByScenarioId(String scenarioId) {
-        return mappingRepository.findByScenarioKey_ScenarioId(scenarioId);
-    }
-
-    /**
-     * 매핑별 테스트케이스 조회
-     */
-    private List<TestCaseEntity> getTestCasesByMappings(List<MappingEntity> mappings) {
-        if (mappings.isEmpty()) {
-            return List.of();
-        }
-        return testCaseRepository.findByMappingKeyIn(mappings);
-    }
-
-    /**
-     * 테스트케이스별 데이터 조회
-     */
-    private List<TestCaseDataEntity> getTestCaseDataByTestCases(List<TestCaseEntity> testCases) {
-        if (testCases.isEmpty()) {
-            return List.of();
-        }
-        return testCaseDataRepository.findByTestcaseKeyIn(testCases);
-    }
-
+    TestCaseReportResponseDto downloadTestCaseReport(String scenarioId, String avalon) throws IOException;
 }
